Add files via upload
Hello Everyone,
The above code is for data encryption and decryption , we have so many applications to check whether our files have been compromised or not for example, HASH CALC is used to check by calculating the SHA and MD5 values of a  particular file, similarly this python code helps us to send our message secretly to the next person without any third parties influence  unless and until a person knows a key they won't be able to decrypt our message. So that we will not become a prey to any of the hackers.
Thank you.